# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HealthFlow
  description: Hospital Appointment System
  version: 1.0.0
servers:
  - url: 'https://localhost:8080'
    description: Local Development Server
paths:

  /patients:
    get:
      tags:
        - Patients
      summary: Returns a list of all patients
      description: Returns all registered patients in the database.
      operationId: listPatients
      responses:
        "200":
          description: A JSON array of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientResponse'
              example:
                - uniqueId: '550e8400-e29b-41d4-a716-446655440000'
                  name: Ahmet
                  surname: Soydaş
                  birthDate: '1986-04-03'
                - uniqueId: '6f9619ff-8b86-d011-b42d-00cf4fc964ff'
                  name: Sefa
                  surname: Soysal
                  birthDate: '2004-08-02'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Patients
      summary: Adds new patient to the database.
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
            example:
              nationalId: '22222222222'
              name: Mehmet
              surname: Yaşaroğlu
              birthDate: '1978-07-23'
              email: mehmet.yasaroglu@gmail.com
              phoneNumber: '+905554123233'
      responses:
        '201':
          description: Patient created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          description: Needed user info is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                code: BAD_REQUEST_ERROR
                message: Needed user info is missing.
                timestamp: '2023-05-14T17:30:00Z'
        '409':
          description: A patient with the same national ID exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 409
                code: PATIENT_ALREADY_EXISTS
                message: A patient with the same national ID already exists.
                timestamp: '2023-05-14T17:30:00Z'
        '500':
          description: Server side error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /patients/{uniqueId}:
    get:
      tags:
        - Patients
      parameters:
        - name: uniqueId
          in: path
          required: true
          description: The unique ID of the patient.
          schema:
            type: string
            format: uuid
      summary: Returns the patient with unique id number.
      description: Returns the patient with unique id number.
      operationId: findPatientWithUniqueId
      responses:
        '200':
          description: Patient with the unique ID returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
              example:
                uniqueId: '550e8400-e29b-41d4-a716-446655440000'
                name: Ahmet
                surname: Soydaş
                birthDate: '1986-04-03'
        '400':
          description: The search request is invalid or the unique ID is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                code: BAD_REQUEST_ERROR
                message: The search request is invalid or the unique ID is missing.
                timestamp: '2023-05-14T17:30:00Z'
        '404':
          description: There is no patient with this ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 404
                code: ID_NOT_FOUND_ERROR
                message: There is no patient with this ID.
                timestamp: '2023-05-14T17:30:00Z'

  /patients/search-by-national-id:
    post:
      tags:
        - Patients
      summary: Returns the patient with matching national ID.
      operationId: findPatientByNationalId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientSearchByNationalIdRequest'
      responses:
        '200':
          description: Patient with the matching national ID successfully returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
              example:
                uniqueId: '550e8400-e29b-41d4-a716-446655440000'
                name: Sefa
                surname: Soysal
                birthDate: '2004-08-02'
        '400':
          description: The search request is invalid or the national ID is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                code: BAD_REQUEST_ERROR
                message: The search request is invalid or the national ID is missing.
                timestamp: '2023-05-14T17:30:00Z'
        '404':
          description: No patient was found with the provided national ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 404
                code: NOT_FOUND_ERROR
                message: No patient was found with the provided national ID.
                timestamp: '2023-05-14T17:30:00Z'

  /patients/search:
    get:
      tags:
        - Patients
      summary: Search patients by name or surname
      operationId: searchPatients
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
            minLength: 1
            example: 'Sefa Soysal'
      responses:
        '200':
          description: Patients matching with the search query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientResponse'
              example:
                - uniqueId: '550e8400-e29b-41d4-a716-446655440000'
                  name: Sefa
                  surname: Soysal
                  birthDate: '2004-08-02'
        '400':
          description: Invalid search request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                code: BAD_REQUEST_ERROR
                message: Invalid search request.
                timestamp: '2023-05-14T17:30:00Z'

components:
  schemas:
    PatientResponse:
      type: object
      required:
        - uniqueId
        - name
        - surname
        - birthDate
      properties:
        uniqueId:
          type: string
          format: uuid
        name:
          type: string
          example: Mehmet
        surname:
          type: string
          example: Yaşaroğlu
        birthDate:
          type: string
          format: date
          example: '1978-07-23'

    ErrorResponse:
      type: object
      required:
        - status
        - code
        - message
        - timestamp
      properties:
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      example:
        status: 500
        code: INTERNAL_SERVER_ERROR
        message: An unexpected server side error occurred.
        timestamp: '2026-09-05T17:30:00Z'

    CreatePatientRequest:
      type: object
      required:
        - nationalId
        - name
        - surname
        - birthDate
        - email
        - phoneNumber
      properties:
        name:
          type: string
        surname:
          type: string
        birthDate:
          type: string
          format: date
        nationalId:
          $ref: '#/components/schemas/NationalId'
        email:
          type: string
          format: email
        phoneNumber:
          type: string

    NationalId:
      type: string
      description: National identification number of the patient.
      example: '22222222222'

    PatientSearchByNationalIdRequest:
      type: object
      required:
        - nationalId
      properties:
        nationalId:
          $ref: '#/components/schemas/NationalId'